--- client/views/settings/menu/menu.html
@@ -1,11 +1,0 @@
-<template name="settingsMenu">
-  <ul class="list-group">
-    <a href="{{ pathFor route='profile' }}" class="list-group-item {{ isActive 'profile' }}">Profile<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
-    <a href="{{ pathFor route='account' }}" class="list-group-item {{ isActive 'account' }}">Account<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
-    {{ # if isInRole 'admin' }}
-      <a href="{{ pathFor route='branding' }}" class="list-group-item {{ isActive 'branding' }}">Branding<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
-    {{ / if }}
-    <a href="{{ pathFor route='settings' }}" class="list-group-item {{ isActive 'settings' }}">Settings<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
-    <a href="{{ pathFor route='settings.proxies' }}" class="list-group-item {{ isActive 'settings.proxies' }}">Proxies<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
-  </ul>
-</template>
