+++ core/client/layouts/master_layout/master_layout.js
@@ -1,4 +1,4 @@
-import { ProjectLogo } from '/logo/collection/collection';
+import { ProjectLogo } from '/settings/logo/collection';
 
 Template.masterLayout.created = function () {
   // Subscription to branding collection
