+++ home/client/header/homeHeader.js
@@ -1,4 +1,4 @@
-import { ProjectLogo } from '/logo/collection/collection';
+import { ProjectLogo } from '/settings/logo/collection';
 
 Template.homeHeader.onCreated(function() {
   const instance = this;
