+++ lib/router/main.js
@@ -1,9 +1,4 @@
-Router.route("/settings", {
-  layout: "masterLayout",
-  template: "settings",
-}, {
-  name: "settings"
-});
+
 
 Router.map(function() {
 
@@ -38,12 +33,6 @@ Router.map(function() {
     path: "/users",
     layoutTemplate: "masterLayout",
     render: "accountsAdmin"
-  });
-
-  this.route("branding", {
-    path: "/branding",
-    layoutTemplate: "masterLayout",
-    render: "branding"
   });
 
   this.route("notAuthorized", {
