+++ lib/router/users.js
@@ -1,17 +1,8 @@
-Router.map(function() {
-  this.route("profile", {
-    path: "/profile"
-  });
-  this.route("account", {
-    path: "/account"
-  });
-
-  this.route('signOut', {
-    path: '/sign-out',
-    onBeforeAction: function() {
-      Meteor.logout(function() {});
-      this.redirect('/');
-      return this.next();
-    }
-  });
+Router.route('signOut', {
+  path: '/sign-out',
+  onBeforeAction: function() {
+    Meteor.logout(function() {});
+    this.redirect('/');
+    return this.next();
+  }
 });
