--- proxies/router.js
@@ -1,4 +1,0 @@
-Router.route("/settings/proxies", {
-  layout: "masterLayout",
-  template: "proxies",
-});
