+++ settings/branding/client/branding.js
