+++ settings/branding/collection/index.js
