+++ settings/logo/collection/server/permissions.js
@@ -1,4 +1,4 @@
-import { ProjectLogo } from '/logo/collection/collection';
+import { ProjectLogo } from '/settings/logo/collection';
 
 ProjectLogo.allow({
   insert: function(userId, file) {
