+++ settings/logo/upload/client/upload.css
