+++ settings/logo/upload/client/upload.html
