+++ settings/logo/upload/client/uploadButton/uploadButton.html
