+++ settings/logo/view/client/view.html
