+++ settings/logo/view/client/view.js
@@ -1,4 +1,4 @@
-import { ProjectLogo } from '/logo/collection/collection';
+import { ProjectLogo } from '/settings/logo/collection';
 
 Template.viewProjectLogo.onCreated(function() {
   const instance = this;
