+++ settings/profile/client/copyButton/copyButton.html
