+++ settings/profile/client/copyButton/copyButton.js
