+++ settings/proxies/client/add/autoform.js
