+++ settings/proxies/client/proxies.js
