+++ settings/proxies/collection/index.js
