+++ settings/proxies/collection/schema.js
