+++ settings/proxies/server/publications.js
