+++ settings/settings/client/menu/menu.html
@@ -1,0 +1,11 @@
+<template name="settingsMenu">
+  <ul class="list-group">
+    <a href="{{ pathFor route='settings.profile' }}" class="list-group-item {{ isActive 'settings.profile' }}">Profile<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
+    <a href="{{ pathFor route='settings.account' }}" class="list-group-item {{ isActive 'settings.account' }}">Account<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
+    {{ # if isInRole 'admin' }}
+      <a href="{{ pathFor route='settings.branding' }}" class="list-group-item {{ isActive 'settings.branding' }}">Branding<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
+    {{ / if }}
+    <a href="{{ pathFor route='settings' }}" class="list-group-item {{ isActive 'settings' }}">Settings<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
+    <a href="{{ pathFor route='settings.proxies' }}" class="list-group-item {{ isActive 'settings.proxies' }}">Proxies<span class="glyphicon glyphicon-chevron-right pull-right"></span></a>
+  </ul>
+</template>
